+++ src/DeepLClientOptions.php
@@ -1,0 +1,12 @@
+<?php
+
+// Copyright 2025 DeepL SE (https://www.deepl.com)
+// Use of this source code is governed by an MIT
+// license that can be found in the LICENSE file.
+
+namespace DeepL;
+
+class DeepLClientOptions extends TranslateTextOptions
+{
+
+}