+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* Added support for the Write API in the client library, the implementation
+  can be found in the `DeepLClient` class. Please refer to the README for usage
+  instructions.
+### Changed
+* The main functionality of the library is now also exposed via the `DeepLClient`
+  class. Please change your code to use this over the `Translator` class whenever
+  convenient.
+
+
 ## [1.10.1] - 2024-11-29
 ### Fixed
 * Fixed the `TextResult` constructor to be compatible with pre-1.10 versions, to
@@ -183,6 +194,7 @@ Stable release.
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-php/compare/v1.10.1...HEAD
 [1.10.1]: https://github.com/DeepLcom/deepl-php/compare/v1.10.0...v1.10.1
 [1.10.0]: https://github.com/DeepLcom/deepl-php/compare/v1.9.0...v1.10.0
 [1.9.0]: https://github.com/DeepLcom/deepl-php/compare/v1.8.0...v1.9.0